rename: 썰 열람페이지 component 분리리
